+++ libuvxx_rtsp/src/rtsp_server.cpp
@@ -8,12 +8,12 @@ rtsp_server::rtsp_server()
 {
 }
 
-int rtsp_server::port() const
+uint16_t rtsp_server::port() const
 {
 	return __rtsp_server_impl ? __rtsp_server_impl->port() : 0;
 }
 
-void rtsp_server::start_server(int port)
+void rtsp_server::start_server(uint16_t port)
 {
     __rtsp_server_impl = std::make_shared<_rtsp_server_impl>(port);
 }
